requirements.txt
blinker==1.6.2
cachetools==5.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.3.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyTelegramBotAPI==4.12.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==60.5.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.3.6
wheel==0.37.1
zipp==3.15.0